[driver] - cleaned up pipeline tests
