+++ driver/pipeline/control_test.cpp
@@ -7,74 +7,12 @@
 // License, use of this software will be governed by the Apache License, Version 2.0,
 // included in the file licenses/APL.txt.
 
-/// GTest
+/// external
 #include "gtest/gtest.h"
 
+/// internal
 #include "driver/pipeline/control.h"
-
-struct MockStreamerConfig {
-    std::shared_ptr<std::vector<synnax::Frame> > reads;
-    std::shared_ptr<std::vector<freighter::Error> > read_errors;
-    freighter::Error close_err;
-};
-
-class MockStreamer final : public pipeline::Streamer {
-public:
-    MockStreamerConfig config;
-    size_t current_read = 0;
-
-    explicit MockStreamer(
-        MockStreamerConfig config
-    ) : config(std::move(config)) {
-    }
-
-    std::pair<synnax::Frame, freighter::Error> read() override {
-        if (current_read >= config.reads->size()) {
-            // block "indefinitely"
-            std::this_thread::sleep_for(std::chrono::milliseconds(5));
-            return {synnax::Frame(0), freighter::NIL};
-        }
-        auto fr = std::move(config.reads->at(current_read));
-        auto err = config.read_errors->at(current_read);
-        current_read++;
-        return {std::move(fr), freighter::NIL};
-    }
-
-    freighter::Error close() override { return config.close_err; }
-};
-
-class MockStreamerFactory final : public pipeline::StreamerFactory {
-public:
-    std::vector<freighter::Error> open_errors;
-    std::shared_ptr<std::vector<MockStreamerConfig> > configs;
-    StreamerConfig config;
-    size_t streamer_opens = 0;
-
-    MockStreamerFactory(
-        std::vector<freighter::Error> open_errors,
-        std::shared_ptr<std::vector<MockStreamerConfig> > configs
-    ) : open_errors(std::move(open_errors)), configs(std::move(configs)) {
-    }
-
-    std::pair<std::unique_ptr<pipeline::Streamer>, freighter::Error> openStreamer(
-        synnax::StreamerConfig config) override {
-        this->streamer_opens++;
-        this->config = config;
-        /// try to grab the next config
-        size_t idx = this->streamer_opens - 1;
-        if (this->streamer_opens > this->configs->size())
-            idx = this->configs->size() - 1;
-        // try to grab the first error. if not, freighter nil
-        auto err = this->streamer_opens > this->open_errors.size()
-                       ? freighter::NIL
-                       : this->open_errors.at(this->streamer_opens - 1);
-        if (err) return {nullptr, err};
-        return {
-            std::make_unique<MockStreamer>((*this->configs)[idx]),
-            freighter::NIL
-        };
-    }
-};
+#include "driver/pipeline/mock/pipeline.h"
 
 class MockSink final : public pipeline::Sink {
 public:
@@ -86,10 +24,10 @@ public:
                  write_errors(std::make_shared<std::vector<freighter::Error> >()) {
     }
 
-    freighter::Error write(synnax::Frame frame) override {
+    freighter::Error write(const synnax::Frame &frame) override {
         if (frame.size() == 0) return freighter::NIL;
-        this->writes->push_back(std::move(frame));
-        // try to grab and remove the firste rror. if not, friehgter nil
+        this->writes->emplace_back(frame.deep_copy());
+        // try to grab and remove the first error. if not, freighter nil
         if (this->write_errors->empty()) return freighter::NIL;
         auto err = this->write_errors->front();
         this->write_errors->erase(this->write_errors->begin());
@@ -103,22 +41,22 @@ public:
 
 TEST(ControlPipeline, testHappyPath) {
     auto fr_1 = synnax::Frame(1);
-    fr_1.add(1, synnax::Series(1.0));
+    fr_1.emplace(1, synnax::Series(1.0));
     auto fr_2 = synnax::Frame(1);
-    fr_2.add(1, synnax::Series(2.0));
-    auto reads = std::make_shared<std::vector<synnax::Frame> >();
+    fr_2.emplace(1, synnax::Series(2.0));
+    const auto reads = std::make_shared<std::vector<synnax::Frame> >();
     reads->push_back(std::move(fr_1));
     reads->push_back(std::move(fr_2));
-    auto read_errors = std::make_shared<std::vector<freighter::Error> >(
-        std::vector<freighter::Error>{
+    const auto read_errors = std::make_shared<std::vector<freighter::Error> >(
+        std::vector{
             freighter::NIL,
             freighter::NIL,
         });
-    auto streamer_config = synnax::StreamerConfig{.channels = {1}};
-    auto streamer_factory = std::make_shared<MockStreamerFactory>(
+    const auto streamer_config = synnax::StreamerConfig{.channels = {1}};
+    const auto streamer_factory = std::make_shared<MockStreamerFactory>(
         std::vector<freighter::Error>{},
         std::make_shared<std::vector<MockStreamerConfig> >(
-            std::vector<MockStreamerConfig>{
+            std::vector{
                 MockStreamerConfig{
                     reads,
                     read_errors,
@@ -126,7 +64,7 @@ TEST(ControlPipeline, testHappyPath) {
                 }
             })
     );
-    auto sink = std::make_shared<MockSink>();
+    const auto sink = std::make_shared<MockSink>();
     auto control = pipeline::Control(
         streamer_factory,
         streamer_config,
@@ -140,13 +78,13 @@ TEST(ControlPipeline, testHappyPath) {
 }
 
 TEST(ControlPipeline, testUnknownErrOnOpen) {
-    auto streamer_factory = std::make_shared<MockStreamerFactory>(
-        std::vector<freighter::Error>{
+    const auto streamer_factory = std::make_shared<MockStreamerFactory>(
+        std::vector{
             freighter::UNKNOWN
         },
         std::make_shared<std::vector<MockStreamerConfig> >()
     );
-    auto sink = std::make_shared<MockSink>();
+    const auto sink = std::make_shared<MockSink>();
     auto control = pipeline::Control(
         streamer_factory,
         StreamerConfig{},
@@ -162,26 +100,26 @@ TEST(ControlPipeline, testUnknownErrOnOpen) {
 
 TEST(ControlPipeline, testOpenRetrySuccessful) {
     auto fr_1 = synnax::Frame(1);
-    fr_1.add(1, synnax::Series(1.0));
+    fr_1.emplace(1, synnax::Series(1.0));
     auto fr_2 = synnax::Frame(1);
-    fr_2.add(1, synnax::Series(2.0));
+    fr_2.emplace(1, synnax::Series(2.0));
     auto reads = std::make_shared<std::vector<synnax::Frame> >();
     reads->push_back(std::move(fr_1));
     reads->push_back(std::move(fr_2));
     auto read_errors = std::make_shared<std::vector<freighter::Error> >(
-        std::vector<freighter::Error>{
+        std::vector{
             freighter::NIL,
             freighter::NIL,
         });
     auto streamer_config = synnax::StreamerConfig{.channels = {1}};
     auto streamer_factory = std::make_shared<MockStreamerFactory>(
-        std::vector<freighter::Error>{
+        std::vector{
             freighter::UNREACHABLE,
             freighter::UNREACHABLE,
             freighter::NIL
         },
         std::make_shared<std::vector<MockStreamerConfig> >(
-            std::vector<MockStreamerConfig>{
+            std::vector{
                 MockStreamerConfig{
                     reads,
                     read_errors,
